feat: move TcpServerConfig and use lock-free elements
